// src/interfaces/players.ts
export interface IPlayer {
  name: string;
  pts: number;
  $: string;
  $$: number;
  id: string;
  mv: number;
  order: number;
  pos: string;
  status: string;
  tee: string;
  thru: number;
  today: string;
  toPar: string;
  tot: number;
  today$: number;
  cut: boolean;
  detail: string;
  r1: number;
  r2: number;
  r3: number;
  r4: number;
  fantasy_points: number;
}